New materials

+++ uzdevumi/celsijs_farenhejts.py
@@ -1,3 +1,3 @@
-celsijs = int(input("Ievadiet grādus Celsija: "))
+celsijs = float(input("Ievadiet grādus Celsija: "))
 farenhejts = celsijs*9/5+32
 print(f"Tas ir {farenhejts} grādu pēc Farenheita skalas.")
